refactor(EditGroup): replace AddGroupAclButton with LCombobox for group selection
